requirements.txt
fastapi
numpy
SQLAlchemy
python-dotenv